AHCI: Use fls() from BitUtils.

+++ src/add-ons/kernel/busses/scsi/ahci/ahci_controller.cpp
@@ -11,6 +11,7 @@
 #include <stdio.h>
 #include <string.h>
 #include <new>
+#include <util/BitUtils.h>
 
 #define TRACE(a...) dprintf("ahci: " a)
 #define FLOW(a...)	dprintf("ahci: " a)
@@ -335,7 +336,8 @@ AHCIController::ResetController()
 	if (fPCIVendorID == PCI_VENDOR_INTEL) {
 		// Intel PCS—Port Control and Status
 		// SATA port enable bits must be set
-		int portCount = std::max(fls(fRegs->pi), 1 + (int)((fRegs->cap >> CAP_NP_SHIFT) & CAP_NP_MASK));
+		int portCount = std::max((int)fls(fRegs->pi),
+			1 + (int)((fRegs->cap >> CAP_NP_SHIFT) & CAP_NP_MASK));
 		if (portCount > 8) {
 			// TODO: fix this when specification available
 			TRACE("don't know how to enable SATA ports 9 to %d\n", portCount);
